Use `Return_Status` enum as return type
This commit refactors the code and makes the `Return_Status` enum be a possible
return type for many revealing functions, instead of using `uint8_t`, making
them more clear and concise.

The possibility `Return_Status__Error` has also been renamed to
`Return_Status__Failure` to avoid making a relation with errors thrown by the
program.

Now, in the main function, when parsing data type and non data type options, the
`Reveal_Entry` function return will be tested against `Return_Status__Failure`,
and not to check if it is zero directly, to make it more readable.