Reduce memory allocation when revealing directory
This commit adds a fix to a very large memory allocation when revealing
directory contents: `malloc` will now consider the right string size for each
directory entry name instead of the 256 bytes pre-allocated in the stack for
that array of characters.

This issue was only detect by `valgrind` when revealing the contents of
directories with large amount of entries.

The version in the binary source code has been bumped to `v9.1.1` to include
this change.